refactor: modularize library structure, support session reuse and async cache checks
