+++ openeihttp/cache.py
@@ -2,11 +2,14 @@
 
 import json
 import logging
-from os.path import dirname, exists, join, split
+from os.path import dirname, join, split
 from typing import Any
 
 import aiofiles
 import aiofiles.os
+import aiofiles.ospath
+
+from .const import MIN_CACHE_SIZE
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -21,9 +24,9 @@ class OpenEICache:
         self._cache_file = cache_file
         self._directory, self._filename = split(cache_file)
 
-    async def write_cache(self, data: Any) -> None:
+    async def write_cache(self, data: bytes) -> None:
         """Write cache file."""
-        if self._directory != "" and not exists(self._directory):
+        if self._directory != "" and not await aiofiles.ospath.exists(self._directory):
             _LOGGER.debug("Directory missing creating: %s", self._directory)
             await aiofiles.os.makedirs(self._directory)
         async with aiofiles.open(self._cache_file, mode="wb") as file:
@@ -33,7 +36,7 @@ class OpenEICache:
     async def read_cache(self) -> Any:
         """Read cache file."""
         _LOGGER.debug("Attempting to read file: %s", self._cache_file)
-        if exists(self._cache_file):
+        if await aiofiles.ospath.exists(self._cache_file):
             async with aiofiles.open(self._cache_file) as file:
                 _LOGGER.debug("Reading file: %s", self._cache_file)
                 value = await file.read()
@@ -53,7 +56,7 @@ class OpenEICache:
         if check:
             size = await aiofiles.os.path.getsize(self._cache_file)
             _LOGGER.debug("Checking cache file size: %s", size)
-            return size > 194
+            return size > MIN_CACHE_SIZE
         return False
 
     async def clear_cache(self) -> None:
